ml/requirements.txt
# ML Training Pipeline Requirements
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
tqdm==4.66.1

# Data collection
yt-dlp==2023.11.16

# Streaming training
spotipy==2.23.0
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4

# Visualization
streamlit==1.28.0
plotly==5.18.0
seaborn==0.13.0
umap-learn==0.5.5
pandas==2.1.4

# Gemini API for code generation
google-generativeai==0.3.2

# YouTube API for trending discovery
google-api-python-client==2.100.0
